Remove most of SVG dependency from platform
