+++ js/platform/document.js
@@ -1,17 +1,14 @@
 import { BasePlatform, LossPlotter } from './base.js'
+import DocumentScatterRenderer from '../renderer/document.js'
 
 export default class DocumentPlatform extends BasePlatform {
 	constructor(task, manager) {
 		super(task, manager)
+		this._renderer.terminate()
+		this._renderer = new DocumentScatterRenderer(manager)
 	}
 
 	init() {
-		if (this.svg.select('g.dc-render').size() === 0) {
-			this.svg.append('g').classed('dc-render', true).style('transform', 'scale(1, -1) translate(0, -100%)')
-		}
-		this._r = this.svg.select('g.dc-render')
-		this._r.selectAll('*').remove()
-
 		this.render()
 		if (this._loss) {
 			this._loss.terminate()
@@ -24,52 +21,19 @@ export default class DocumentPlatform extends BasePlatform {
 	}
 
 	set trainResult(value) {
-		this._pred = value
-		this._displayResults(value, this.trainInput)
+		this._renderer.trainResult = value
 	}
 
 	testInput() {
-		const x = this.datas.x[0]
-		const [words, idxs] = this.datas.ordinal(x)
-		return words
+		return this._renderer.testData()
 	}
 
 	testResult(value) {
-		this._pred = value
-		this._displayResults(value, this.testInput())
+		this._renderer.testResult(value)
 	}
 
-	render() {}
-
-	_displayResults(data, words) {
-		this._r.selectAll('*').remove()
-		let y_max = []
-		let y_min = []
-		for (let i = 0; i < data[0].length; i++) {
-			const ym = data.map(v => v[i])
-			y_max.push(Math.max(...ym))
-			y_min.push(Math.min(...ym))
-		}
-
-		const width = this.width
-		const height = this.height - 20
-
-		const scales = [width, height].map((m, i) => (m - 10) / (y_max[i] - y_min[i]))
-		const scale_min = Math.min(...scales)
-		const offsets = [5, 20]
-		for (let i = 0; i < scales.length; i++) {
-			if (scales[i] > scale_min) {
-				if (!isFinite(scales[i])) {
-					offsets[i] = [width, height][i] / 2 - y_min[i]
-				} else {
-					offsets[i] += ((scales[i] - scale_min) * (y_max[i] - y_min[i])) / 2
-				}
-			}
-		}
-		for (let i = 0; i < data.length; i++) {
-			const v = data[i].map((a, k) => (a - y_min[k]) * scale_min + offsets[k])
-			this._r.append('text').attr('x', v[0]).attr('y', v[1]).text(words[i]).append('title').text(words[i])
-		}
+	render() {
+		this._renderer.render()
 	}
 
 	plotLoss(value) {
@@ -80,7 +44,6 @@ export default class DocumentPlatform extends BasePlatform {
 	}
 
 	terminate() {
-		this._r.remove()
 		if (this._loss) {
 			this._loss.terminate()
 		}
