+++ js/renderer/scatter.js
@@ -1,5 +1,6 @@
 import BaseRenderer from './base.js'
-import { getCategoryColor, DataPoint, DataCircle, DataHulls } from '../utils.js'
+import { getCategoryColor, specialCategory, DataPoint, DataCircle, DataLine, DataHulls } from '../utils.js'
+import Matrix from '../../lib/util/matrix.js'
 
 const scale = function (v, smin, smax, dmin, dmax) {
 	if (!isFinite(smin) || !isFinite(smax) || smin === smax) {
@@ -96,9 +97,127 @@ export default class ScatterRenderer extends BaseRenderer {
 		return this._p
 	}
 
+	set trainResult(value) {
+		const task = this._manager.platform.task
+		if (task === 'AD') {
+			if (this._svg.select('.tile').size() === 0) {
+				this._svg.insert('g', ':first-child').classed('tile', true).classed('anormal_point', true)
+			}
+			const r = this._svg.select('.tile')
+			r.selectAll('*').remove()
+			value.forEach((v, i) => {
+				if (v) {
+					const o = new DataCircle(r, this.points[i])
+					o.color = getCategoryColor(specialCategory.error)
+				}
+			})
+		} else if (task === 'SC') {
+			if (this._svg.select('.tile').size() === 0) {
+				this._svg.insert('g', ':first-child').classed('tile', true)
+			}
+			const r = this._svg.select('.tile')
+			r.selectAll('*').remove()
+
+			value.forEach((v, i) => {
+				const o = new DataCircle(r, this.points[i])
+				o.color = getCategoryColor(v)
+			})
+		} else if (task === 'DR' || task === 'FS' || task === 'TF') {
+			if (this._svg.select('.tile').size() === 0) {
+				this._svg.insert('g', ':first-child').classed('tile', true).attr('opacity', 0.5)
+			}
+			const r = this._svg.select('.tile')
+			r.selectAll('*').remove()
+
+			const d = value[0].length
+			let y = value
+			if (d === 1) {
+				y = y.map(v => [v, 0])
+			}
+			let y_max = []
+			let y_min = []
+			for (let i = 0; i < y[0].length; i++) {
+				const ym = y.map(v => v[i])
+				y_max.push(Math.max(...ym))
+				y_min.push(Math.min(...ym))
+			}
+
+			const ranges = this.datas.dimension <= 1 ? [this.height, this.height] : [this.width, this.height]
+
+			const scales = ranges.map((m, i) => (m - 10) / (y_max[i] - y_min[i]))
+			let scale_min = Math.min(...scales)
+			const offsets = [5, 5]
+			for (let i = 0; i < scales.length; i++) {
+				if (!isFinite(scale_min) || scales[i] > scale_min) {
+					if (!isFinite(scales[i])) {
+						offsets[i] = ranges[i] / 2 - y_min[i]
+					} else {
+						offsets[i] += ((scales[i] - scale_min) * (y_max[i] - y_min[i])) / 2
+					}
+				}
+			}
+			if (!isFinite(scale_min)) {
+				scale_min = 0
+			}
+
+			let min_cost = Infinity
+			let min_cost_y = null
+			const p = Matrix.fromArray(this.points.map(p => p.at))
+			for (let i = 0; i < (this.datas.dimension <= 1 ? 1 : 2 ** d); i++) {
+				const rev = i
+					.toString(2)
+					.padStart(d, '0')
+					.split('')
+					.map(v => !!+v)
+
+				const ry = y.map(v => {
+					return v.map((a, k) => ((rev[k] ? y_max[k] - a + y_min[k] : a) - y_min[k]) * scale_min + offsets[k])
+				})
+				const y_mat = Matrix.fromArray(ry)
+				y_mat.sub(p)
+				const cost = y_mat.norm()
+				if (cost < min_cost) {
+					min_cost = cost
+					min_cost_y = ry
+				}
+			}
+
+			min_cost_y.forEach((v, i) => {
+				const p = new DataPoint(
+					r,
+					this.datas.dimension <= 1 ? [this.points[i].at[0], v[0]] : v,
+					this.points[i].category
+				)
+				p.radius = 2
+				const dl = new DataLine(r, this.points[i], p)
+				dl.setRemoveListener(() => p.remove())
+			})
+		} else if (task === 'GR') {
+			if (this._svg.select('.tile').size() === 0) {
+				this._svg
+					.insert('g', ':first-child')
+					.classed('tile', true)
+					.classed('generated', true)
+					.attr('opacity', 0.5)
+			}
+			const r = this._svg.select('.tile.generated')
+			r.selectAll('*').remove()
+			let cond = null
+			if (Array.isArray(value) && value.length === 2 && Array.isArray(value[0]) && Array.isArray(value[0][0])) {
+				;[value, cond] = value
+			}
+
+			value.forEach((v, i) => {
+				let p = new DataPoint(r, this.toPoint(v), cond ? cond[i][0] : 0)
+				p.radius = 2
+			})
+		}
+	}
+
 	init() {
 		this._lastpred = null
 		this._r_tile?.remove()
+		this._svg.select('.tile').remove()
 		this._make_selector()
 	}
